[refactor] 헬퍼 서비스를 위한 어노테이션 추가 (#3)
