+++ src/main/java/konkuk/chacall/global/common/annotation/HelperService.java
@@ -1,0 +1,14 @@
+package konkuk.chacall.global.common.annotation;
+
+import org.springframework.stereotype.Service;
+
+import java.lang.annotation.ElementType;
+import java.lang.annotation.Retention;
+import java.lang.annotation.RetentionPolicy;
+import java.lang.annotation.Target;
+
+@Target({ElementType.TYPE})
+@Retention(RetentionPolicy.RUNTIME)
+@Service
+public @interface HelperService {
+}
